Modify Regression Table Labels and Layout Allows customization of publication ready tables Designed for use with tables generated by regression functions

Modify Regression Table Labels and Layout Allows customization of publication ready tables Designed for use with tables generated by regression functions

Usage

modify_table(
  gt_table,
  variable_labels = NULL,
  level_labels = NULL,
  header_labels = NULL,
  caption = NULL,
  bold_labels = FALSE,
  bold_levels = FALSE,
  remove_N = FALSE,
  remove_N_obs = FALSE,
  remove_abbreviations = FALSE,
  caveat = NULL
)

Arguments

gt_table

A `gtsummary` table object

variable_labels

A named vector to relabel variable names.

level_labels

A named vector to relabel levels of variables.

header_labels

A named vector to relabel table column headers. Names must match internal column names (e.g., `"estimate"`, `"p.value"`).

caption

A character string to add a caption (table title).

bold_labels

Logical. If `TRUE`, variable labels are bolded.

bold_levels

Logical. If `TRUE`, factor level labels are bolded.

remove_N

Logical. If `TRUE`, hides the sample size column (`N`) in univariate regression tables (`uni_reg`, `uni_reg_nbin`). Ignored for multivariate tables with a warning.

remove_N_obs

Logical. If `TRUE`, removes source note showing N_obs sample size in multivariate regression tables (`multi_reg`, `multi_reg_nbin`). Ignored for univariate tables with a warning.

remove_abbreviations

Logical. If `TRUE`, removes abbreviations showing expansions of effect size and confidence interval for all regression table outputs

caveat

A character string to add as a source note below the table (e.g., "N may vary due to missing data."). This is always applied regardless of table type.

Value

A customized `gtsummary` table object with modified labels, layout, and options.

Examples

if (FALSE) { # \dontrun{
library(gtregression)
tbl_custom <- modify_table(
  uni_rr,
  variable_labels = c(age_cat = "Age", bmi = "BMI"),
  level_labels = list(age_cat = c(`Young` = "Young Adults",
  `Old` = "Older Adults")),
  header_labels = c(estimate = "**Risk Ratio**",
  `p.value` = "***P*-value**"),
  caption = "Table 1: Univariate Regression",
  bold_labels = TRUE,
  remove_N = TRUE,
  caveat = "N may vary due to missing data."
)
} # }