Visualize a Regression Model as a Forest Plot

Creates a forest plot from a `gtsummary` object. Supports both univariate and multivariable models with hierarchical labels for categorical variables. Designed to work seamlessly with outputs from functions like `uni_reg()` and `multi_reg()`.

Usage

plot_reg(
  tbl,
  title = NULL,
  ref_line = 1,
  order_y = NULL,
  log_x = FALSE,
  xlim = NULL,
  breaks = NULL,
  point_color = "#1F77B4",
  errorbar_color = "#4C4C4C",
  base_size = 14,
  show_ref = TRUE
)

Arguments

tbl

A `gtsummary` object from regression functions (e.g., `uni_reg`, `multi_reg`).

title

Optional plot title (character).

ref_line

Numeric value for the reference line (default = 1).

order_y

Optional character vector specifying the custom order of the y-axis labels.

log_x

Logical. If `TRUE`, uses a logarithmic x-axis (default = FALSE).

xlim

Optional numeric vector specifying x-axis limits (e.g., `c(0.1, 10)`).

breaks

Optional numeric vector for x-axis tick breaks.

point_color

Color of the points (default is automatic).

errorbar_color

Color of the error bars (default is automatic).

base_size

Base font size for text elements.

show_ref

Logical. If `TRUE`, includes reference categories in the plot (default = TRUE).

Value

A `ggplot2` object representing the forest plot.

Examples

if (FALSE) { # \dontrun{
data <- data_PimaIndiansDiabetes
uni_model <- uni_reg(data, "glucose", "bmi", approach = "linear")
plot_reg(uni_model)
} # }